Display welcome screen on Duplicates page (#54)
